requirements.txt
Appium-Python-Client==0.49
behave==1.2.6
certifi==2019.11.28
decorator==4.4.1
parse==1.14.0
parse-type==0.5.2
pipenv==2018.11.26
public==2019.4.13
selenium==3.141.0
self==2019.4.13
six==1.14.0
urllib3==1.25.8
PySocks==1.7.1
mysql-connector-python==8.0.19
distlib==0.3.0
filelock==3.0.12
parse==1.14.0
pymongo==3.10.1
urllib3==1.25.8
requests==2.23.0
DateTime==4.3